+++ src/Soluble/Japha/Bridge/Adapter.php
@@ -50,6 +50,7 @@ class Adapter
      * $ba = new Adapter([
      *     'driver' => 'Pjb62',
      *     'servlet_address' => 'http://127.0.0.1:8080/javabridge-bundle/java/servlet.phpjavabridge'
+     *      //'use_persistent_connection' => false
      *      //'java_default_timezone' => null,
      *      //'java_prefer_values' => true,
      *      //'force_simple_xml_parser' => false